CSV download

+++ schemas.py
@@ -1,4 +1,5 @@
 import time
+from decimal import Decimal
 from datetime import datetime
 from typing import Optional
 from pydantic import BaseModel, Field
@@ -24,11 +25,11 @@ class ExpenseResponse(BaseModel):
     createdAt: str = Field(..., description="ISO 8601 UTC timestamp string when the record was created")
     createdAtEpoch: int = Field(..., description="UNIX Epoch time in seconds when the record was created")
     description: Optional[str] = Field(None, description="Optional description of the expense")
-    amount: float = Field(..., description="Amount of the expense", gt=0)
+    amount: Decimal = Field(..., description="Amount of the expense", gt=0)
 
 
 class ExpenseCreate(BaseModel):
     name: str = Field(..., description="Name of the expense", min_length=1)
     categoryId: str = Field(..., description="Identifier of the category")
     description: Optional[str] = Field(None, description="Optional description of the expense")
-    amount: float = Field(..., description="Amount of the expense", gt=0)
+    amount: Decimal = Field(..., description="Amount of the expense", gt=0)
